requirements.txt
langchain
langgraph
langchain-community
langchain-openai
chromadb
python-dotenv
jupyter
pypdf
tiktoken
